Updated AzureContentSafety
