+++ src/TagzApp.Storage.Postgres/PostgresMessagingService.cs
@@ -3,6 +3,7 @@ using Microsoft.Extensions.Caching.Memory;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using TagzApp.Communication;
+using TagzApp.Storage.Postgres.SafetyModeration;
 using TagzApp.Web.Services;
 
 namespace TagzApp.Storage.Postgres;
